Change: illust id from INT to LONG Fixed: DownloadManager filter & sort bug Fixed: some minor bugs

+++ PixivWPF/Common/DownloadItemControl.xaml.cs
@@ -309,7 +309,7 @@ namespace PixivWPF.Common
             disposed = true;
         }
 
-        public void UpdateDownloadState(int? illustid = null, bool? exists = null)
+        public void UpdateDownloadState(long? illustid = null, bool? exists = null)
         {
             if (ProgressPercent == 100)
             {
