+++ PixivWPF/Pages/HistoryPage.xaml.cs
@@ -33,12 +33,12 @@ namespace PixivWPF.Pages
 
         public string Contents { get; set; } = string.Empty;
 
-        private void UpdateDownloadState(int? illustid = null, bool? exists = null)
+        private void UpdateDownloadState(long? illustid = null, bool? exists = null)
         {
             HistoryItems.UpdateDownloadStateAsync(illustid, exists);
         }
 
-        public async void UpdateDownloadStateAsync(int? illustid = null, bool? exists = false)
+        public async void UpdateDownloadStateAsync(long? illustid = null, bool? exists = false)
         {
             await Task.Run(() =>
             {
@@ -46,7 +46,7 @@ namespace PixivWPF.Pages
             });
         }
 
-        public async void UpdateLikeStateAsync(int illustid = -1, bool is_user = false)
+        public async void UpdateLikeStateAsync(long illustid = -1, bool is_user = false)
         {
             await new Action(() =>
             {
@@ -54,7 +54,7 @@ namespace PixivWPF.Pages
             }).InvokeAsync();
         }
 
-        public void UpdateLikeState(int illustid = -1, bool is_user = false)
+        public void UpdateLikeState(long illustid = -1, bool is_user = false)
         {
             HistoryItems.UpdateLikeState(illustid, is_user);
         }
